requirements.txt
beautifulsoup4
fastcore
ghapi
howfairis
jupyterlab
matplotlib
openpyxl
pandas
paperswithcode-client
python-dotenv
requests
rispy
scikit_learn
scipy
seaborn
simplejson